// apps/service-web/src/.generated/api/model/supportMessageItem.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Service API
 * Data Plane Service API Service
 * OpenAPI spec version: 1.0.0
 */
import type { SupportMessageItemReadAt } from './supportMessageItemReadAt';
import type { SupportMessageItemSenderType } from './supportMessageItemSenderType';
import type { SupportMessageItemSenderUserId } from './supportMessageItemSenderUserId';

export interface SupportMessageItem {
  id: string;
  roomId: string;
  /** @nullable */
  senderUserId?: SupportMessageItemSenderUserId;
  senderName: string;
  senderType: SupportMessageItemSenderType;
  content: string;
  /** @nullable */
  readAt?: SupportMessageItemReadAt;
  createdAt: string;
}