// apps/service-web/src/.generated/api/model/supportRoomItem.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Service API
 * Data Plane Service API Service
 * OpenAPI spec version: 1.0.0
 */
import type { SupportRoomItemAssigneeName } from './supportRoomItemAssigneeName';
import type { SupportRoomItemLastMessageAt } from './supportRoomItemLastMessageAt';
import type { SupportRoomItemStatus } from './supportRoomItemStatus';

export interface SupportRoomItem {
  id: string;
  title: string;
  status: SupportRoomItemStatus;
  userId: string;
  userName: string;
  /** @nullable */
  assigneeName?: SupportRoomItemAssigneeName;
  /** @nullable */
  lastMessageAt?: SupportRoomItemLastMessageAt;
  createdAt: string;
  updatedAt: string;
}